♻️[Refactor] 공연 상세 라우트 [id]로 변경 및 ShowDetailScreen 리팩토링
- pages/shows/[id] (폴더 방식으로 변경)
- ShowDetailScreen: props 제거, useRouter + fetch 내부 처리
- ShowDetailScreen: 무한스크롤 / RECORDS_PER_PAGE 제거 (단건 조회)
- isLoading = !isReady || detailLoading 통합으로 스켈레톤 처리 개선
