+++ pages/shows/[id]/index.tsx
@@ -1,0 +1,7 @@
+import ShowDetailScreen from "@/components/features/shows/detail/ShowDetailScreen";
+
+export default function ShowDetailPage() {
+  return <ShowDetailScreen />;
+}
+
+ShowDetailPage.noBottomNav = true;
